Update Index index.adoc
Gif removed and removed the image of the tea cup. Maybe try use the file on GitHub instead of locally? At this point I don't know why it is not building on your side. If you can see something that is wrong in the code maybe just tell me and I remove it.